refactor: rewrite form with redux
